--- src/main/java/com/TwinStar/TwinStar/comment/dto/CommentRequestDto.java
@@ -1,12 +1,0 @@
-package com.TwinStar.TwinStar.comment.dto;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CommentRequestDto {
-    private String content;
-    private Long userId;
-}
