--- src/main/java/com/TwinStar/TwinStar/comment/service/CommentService.java
@@ -1,83 +1,0 @@
-package com.TwinStar.TwinStar.comment.service;
-
-import com.TwinStar.TwinStar.comment.domain.Comment;
-import com.TwinStar.TwinStar.comment.dto.CommentRequestDto;
-import com.TwinStar.TwinStar.comment.dto.CommentResponseDto;
-import com.TwinStar.TwinStar.comment.repository.CommentRepository;
-import com.TwinStar.TwinStar.post.domain.Post;
-import com.TwinStar.TwinStar.post.repository.PostRepository;
-import com.TwinStar.TwinStar.user.domain.User;
-import com.TwinStar.TwinStar.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class CommentService {
-    private final CommentRepository commentRepository;
-    private final PostRepository postRepository;
-    private final UserRepository userRepository;
-
-    @Transactional
-    public CommentResponseDto createComment(Long postId, CommentRequestDto requestDto) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("게시물을 찾을 수 없습니다."));
-        User user = userRepository.findById(requestDto.getUserId())
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-        if (requestDto.getContent().trim().isEmpty()) {
-            throw new IllegalArgumentException("댓글 내용을 입력해주세요.");
-        }
-        Comment comment = Comment.builder()
-                .post(post)
-                .user(user)
-                .content(requestDto.getContent())
-                .delYn("N")
-                .build();
-
-        return CommentResponseDto.fromEntity(commentRepository.save(comment));
-    }
-
-    public List<CommentResponseDto> getComments(Long postId) {
-        return commentRepository.findByPostIdOrderByCreatedTimeDesc(postId).stream()
-                .map(CommentResponseDto::fromEntity)
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public CommentResponseDto updateComment(Long commentId, CommentRequestDto requestDto, Long currentUserId) {
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다."));
-
-        // 댓글 작성자 확인
-        if (!comment.getUser().getId().equals(currentUserId)) {
-            throw new IllegalArgumentException("댓글을 수정할 권한이 없습니다.");
-        }
-
-        comment.updateContent(requestDto.getContent());
-        commentRepository.save(comment); // 명시적으로 저장 (안 해도 변경 감지는 되지만 확실하게 처리)
-
-        return CommentResponseDto.fromEntity(comment);
-    }
-
-
-    // 댓글 삭제
-    @Transactional
-    public void deleteComment(Long commentId) {
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
-
-        comment.deleteComment();
-    }
-
-    @Transactional
-    public List<CommentResponseDto> getCommentsByPostId(Long postId) {
-        List<Comment> comments = commentRepository.findAllCommentsByPostId(postId);
-        return comments.stream()
-                .map(CommentResponseDto::fromEntity)
-                .collect(Collectors.toList());
-    }
-}
