--- src/main/java/com/TwinStar/TwinStar/post/domain/PostStatus.java
@@ -1,6 +1,0 @@
-package com.TwinStar.TwinStar.post.domain;
-
-public enum PostStatus {
-    ACTIVE,
-    DELETE
-}
