--- src/main/java/com/TwinStar/TwinStar/post/dto/PostResponseDto.java
@@ -1,32 +1,0 @@
-package com.TwinStar.TwinStar.post.dto;
-
-import com.TwinStar.TwinStar.post.domain.Post;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PostResponseDto {
-    private Long id;
-    private String content;
-    private String nickName;
-    private String profileImg;
-    private int likesCount;
-    private int commentsCount;
-    private LocalDateTime createdTime;
-
-    public static PostResponseDto fromEntity(Post post) {
-        return PostResponseDto.builder()
-                .id(post.getId())
-                .content(post.getContent())
-                .nickName(post.getUser().getNickName())
-                .profileImg(post.getUser().getProfileImg())
-                .likesCount(post.getScore())
-                .commentsCount(post.getComments().size())
-                .createdTime(post.getCreatedTime())
-                .build();
-    }
-}
