+++ src/main/java/com/TwinStar/TwinStar/post/dto/ProfilePostResDto.java
@@ -1,19 +1,17 @@
 package com.TwinStar.TwinStar.post.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-@Builder
 public class ProfilePostResDto {
-    private Long id;
-    private MultipartFile file;
-    private int postLikeCount;
-    private int commentCount;
-    private int shareCount;
+    private Long postId;
+    private String imageUrl;
+    private Long likeCount;
+    private Long commentCount;
 }
