+++ src/main/java/com/TwinStar/TwinStar/post/service/PostService.java
@@ -1,0 +1,11 @@
+package com.TwinStar.TwinStar.post.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class PostService {
+
+
+}
