--- src/main/java/com/TwinStar/TwinStar/postLike/PostLikeRepository.java
@@ -1,21 +1,0 @@
-package com.TwinStar.TwinStar.postLike;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
-
-    @Query("SELECT pl FROM PostLike pl WHERE pl.post.id = :postId AND pl.user.id = :userId")
-    Optional<PostLike> findByPostIdAndUserId(@Param("postId") Long postId, @Param("userId") Long userId);
-
-    @Query("SELECT COUNT(pl) FROM PostLike pl WHERE pl.post.id = :postId")
-    int countLikesByPostId(@Param("postId") Long postId);
-
-    boolean existsByPostIdAndUserId(Long postId, Long userId);
-}
-
