--- src/main/java/com/TwinStar/TwinStar/postLike/PostLikeService.java
@@ -1,36 +1,0 @@
-package com.TwinStar.TwinStar.postLike;
-
-import com.TwinStar.TwinStar.post.domain.Post;
-import com.TwinStar.TwinStar.post.repository.PostRepository;
-import com.TwinStar.TwinStar.user.domain.User;
-import com.TwinStar.TwinStar.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class PostLikeService {
-
-    private final PostLikeRepository postLikeRepository;
-    private final PostRepository postRepository;
-    private final UserRepository userRepository;
-
-    @Transactional
-    public void likePost(Long postId, Long userId) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("게시물을 찾을 수 없습니다."));
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-        Optional<PostLike> existingLike = postLikeRepository.findByPostIdAndUserId(postId, userId);
-
-        if (existingLike.isPresent()) {
-            postLikeRepository.delete(existingLike.get()); // 좋아요 취소
-        } else {
-            postLikeRepository.save(PostLike.create(post, user)); // 좋아요 추가
-        }
-    }
-}
